Part_2/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backoff==2.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chex==0.1.5
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
dill==0.3.6
dm-tree==0.1.7
entrypoints==0.4
etils==0.8.0
executing==1.1.1
fonttools==4.38.0
frozenlist==1.3.1
icontract==2.6.2
idna==3.4
importlib-resources==5.10.0
ipykernel==6.16.2
ipython==8.5.0
jax==0.3.23
jaxlib==0.3.22
jedi==0.18.1
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
lz4==4.0.2
matplotlib==3.6.1
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.2
nest-asyncio==1.5.6
networkx==2.8.7
ntlm-auth==1.5.0
numpy==1.23.4
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
parso==0.8.3
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
ply==3.11
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
pyzmq==24.0.1
qcware-quasar==1.0.7
qiskit==0.39.0
qiskit-aer==0.11.0
qiskit-ibmq-provider==0.19.2
qiskit-terra==0.22.0
QuantLib==1.28
requests==2.28.1
requests-ntlm==1.1.0
retworkx==0.12.0
rich==12.6.0
rustworkx==0.12.0
scikit-learn==1.1.2
scipy==1.9.3
six==1.16.0
sklearn==0.0
sortedcontainers==2.4.0
stack-data==0.5.1
stevedore==4.1.0
symengine==0.9.2
sympy==1.11.1
tabulate==0.9.0
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
tweedledum==1.1.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
websocket-client==1.4.1
websockets==10.4
yarl==1.8.1
zipp==3.10.0
pytket==1.7.3
pytket-quantinuum==0.9.0
